Elimin clave secreta de SendGrid y actualic .gitignore

+++ adoptapp/src/main/java/com/pruebas/repository/TiendaRepository.java
@@ -12,5 +12,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TiendaRepository extends JpaRepository<TiendaModel, Long> {
+public interface TiendaRepository extends JpaRepository<TiendaModel, Integer> {
 }
